feat(DetailsListItem.tsx): render HTML entities and links in details
This change allows the DetailsListItem component to render HTML
entities and links in the details content. The component now uses
the parseFormattedStringToDomElements function to parse the content
and render it as HTML. This allows users to include HTML entities
and links in the details content, which can improve the user
experience.
The type of content prop is changed to ReactNode to allow JSX
elements to be passed as content.
fp-ts library is used to handle the parsing of the content.