fix(summaryPage.spec.ts): fix summary page e2e tests
The summary page e2e tests were failing due to changes in the data returned by the GraphQL API. This commit updates the tests to use a different gene (DDB_G0269114) and updates the expected values to match the data returned by the API for that gene. Also, the tests were updated to check for the correct number of publications.