feat(goaPage.spec.ts): add e2e tests for Gene Ontology Annotations page
This commit adds end-to-end tests for the Gene Ontology
Annotations page. The tests verify that the page renders
the Molecular Function panel, collapses the panel when the
Accordion Summary is clicked, and sorts data when a property
in the header is clicked. It also tests the API by checking
if the data from graphql endpoint contains the expected values.