fix(genome-page): fix reference page e2e tests
The reference page e2e tests were failing because the date format
was incorrect. The tests were also not robust enough to handle
different date formats. This commit fixes the tests by using the
formatDate function to format the date and by using regular
expressions to match the date in the table.
The test "Clicking on mentioned gene navigates to that gene's summary page"
was also failing due to a typo in the test name. This commit fixes
the typo.