feat(phenotypesPage.spec.ts): add e2e tests for phenotypes page
Adds end-to-end tests for the phenotypes page to ensure that:
- All strains with phenotypes are displayed.
- All phenotypes of a strain are displayed.
- Strains without phenotypes are not displayed.

The tests use Playwright to interact with the page and verify the
presence and absence of specific elements. The tests also use fp-ts
to improve the structure and readability of the assertions.
The eslint rule sonarjs/no-duplicate-string is disabled to allow
duplicate strings in the tests.
The page.goto is changed to `/gene/${TEST_GENE}/phenotypes` to test the
phenotypes page.