feat: upgrade vite to v7 and replace @vitejs/plugin-react with @vitejs/plugin-react-swc
The vite package is upgraded to the latest major version 7.x.
The @vitejs/plugin-react package is replaced with @vitejs/plugin-react-swc.
The vite-plugin-istanbul and vite-plugin-swc-react-refresh packages are removed.
This change is done to take advantage of the latest features and performance improvements in Vite.
The vite-plugin-react-swc plugin is a faster alternative to @vitejs/plugin-react.
The vite-plugin-istanbul plugin is removed as it is no longer needed.
The vite-plugin-swc-react-refresh plugin is removed as it is no longer needed.