feat(PhoneNumberInput.tsx): add phone number validation
Adds phone number validation to the PhoneNumberInput component.
The validation checks if the phone number is valid based on the
selected country code. If the phone number is invalid, a message
is added to the additionalInformation field. If the phone number
is valid, the message is removed from the additionalInformation
field.
The validation is triggered on blur of the phone number input.

The phone number validation is done using the isPhoneValid function
from the utils/checkPhoneValidity file.
The additionalInformation field is updated using the setValue
function from the react-hook-form library.
The INVALID_PHONE_MESSAGE constant is used to store the message
that is added to the additionalInformation field.
The appendIfEmpty function is used to add the message to the
additionalInformation field only if it is not already present.
The Sreplace function is used to remove the message from the
additionalInformation field.
The Sincludes function is used to check if the message is already
present in the additionalInformation field.
The BMonoidAny.concat function is used to combine the results of
the isPhoneValid and SisEmpty functions.
The match function from the ts-pattern library is used to handle
the different cases based on the phone number and comments values.
The eslint-disable comment is added to disable the
@typescript-eslint/no-shadow rule for the file.

This change improves the user experience by providing feedback
on the validity of the phone number.