feat(ui-frontpage): migrate material-ui to mui in ui-frontpage package
The ui-frontpage package was migrated from material-ui to mui. The makeStyles and withStyles functions were replaced with the styled function from @mui/material/styles. The @material-ui/core and @material-ui/icons imports were replaced with @mui/material and @mui/icons-material imports, respectively. The dictyTheme file was updated to use ThemeOptions from @mui/material/styles instead of @material-ui/core. The overrides property in dictyTheme was replaced with the components property and the styleOverrides property.